refactor: extract modal from project modal into reusable component
